renamed probnum.typing to probnum.types
